合并 v1.9

+++ backend/src/main/java/io/metersphere/api/service/ApiScenarioModuleService.java
@@ -190,18 +190,18 @@ public class ApiScenarioModuleService extends NodeTreeService<ApiScenarioModuleD
         request.setUpdateTime(System.currentTimeMillis());
         checkApiScenarioModuleExist(request);
         List<ApiScenarioDTO> apiScenarios = queryByModuleIds(request);
-
         apiScenarios.forEach(apiScenario -> {
             StringBuilder path = new StringBuilder(apiScenario.getModulePath());
             List<String> pathLists = Arrays.asList(path.toString().split("/"));
-            pathLists.set(request.getLevel(), request.getName());
-            path.delete(0, path.length());
-            for (int i = 1; i < pathLists.size(); i++) {
-                path.append("/").append(pathLists.get(i));
+            if (pathLists.size() > request.getLevel()) {
+                pathLists.set(request.getLevel(), request.getName());
+                path.delete(0, path.length());
+                for (int i = 1; i < pathLists.size(); i++) {
+                    path.append("/").append(pathLists.get(i));
+                }
+                apiScenario.setModulePath(path.toString());
             }
-            apiScenario.setModulePath(path.toString());
         });
-
         batchUpdateApiScenario(apiScenarios);
 
         return apiScenarioModuleMapper.updateByPrimaryKeySelective(request);
